# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API server for asset management system.",
        "title": "Asset Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:6500",
    "paths": {
...
        "/base/tax-information/create": {
            "post": {
                "description": "Create one Tax Information record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tax Information"
                ],
                "summary": "Create one Tax Information record",
                "parameters": [
                    {
                        "description": "Tax Information Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TaxInformations"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralCreateResponseBody"
                        }
                    }
                }
            }
        },
        "/base/tax-information/one/{id}": {
            "get": {
                "description": "Get one tax information record by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tax Information"
                ],
                "summary": "Get one tax information record by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tax information ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TaxInformations"
                        }
                    }
                }
            }
        },
...
        "/sys/role/create": {
            "post": {
                "description": "Create one sys role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sys Role"
                ],
                "summary": "Create one sys role",
                "parameters": [
                    {
                        "description": "List Action Record Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SysRoles"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SysRoles"
                        }
                    }
                }
            }
        },
        "/sys/role/one/{id}": {
            "get": {
                "description": "Get one sys role record by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sys Role"
                ],
                "summary": "Get one sys role record by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "sys role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SysRoles"
                        }
                    }
                }
            }
        },
        "/sys/role/update": {
            "post": {
                "description": "Update sys role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sys Role"
                ],
                "summary": "Update sys role",
                "parameters": [
                    {
                        "description": "Update Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SysRoles"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SysRoles"
                        }
                    }
                }
            }
        },
        "/sys/role/void/{id}": {
            "delete": {
                "description": "Delete one sys role record by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sys Role"
                ],
                "summary": "Delete one sys role record by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sys Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralUpdateInactiveResponseBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.GeneralCreateResponseBody": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "dto.GeneralUpdateInactiveResponseBody": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.GeneralUpdateResponse"
                },
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.GeneralUpdateResponse": {
            "type": "object",
            "properties": {
                "MatchedCount": {
                    "type": "integer"
                },
                "ModifiedCount": {
                    "type": "integer"
                },
                "UpsertedCount": {
                    "type": "integer"
                },
                "UpsertedID": {
                    "type": "string"
                }
            }
        },
...
        "models.SysRoles": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "delete": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "menuIds": {
                    "type": "array",
                    "items": {}
                },
                "name": {
                    "type": "string"
                },
                "read": {
                    "type": "boolean"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "update": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                },
                "upload": {
                    "type": "boolean"
                },
                "write": {
                    "type": "boolean"
                }
            }
        },
...
        "models.TaxInformations": {
            "type": "object",
            "properties": {
                "countryCode": {
                    "type": "string"
                },
                "countryName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "importRate": {
                    "type": "number"
                },
                "nationCode": {
                    "type": "string"
                },
                "nationName": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "taxCode": {
                    "type": "string"
                },
                "taxName": {
                    "type": "string"
                },
                "taxRate": {
                    "type": "number"
                },
                "taxType": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}